Rename PyCitySchools_Challenge_Testing.ipynb to PyCitySchools_with_bugs.ipynb
My first run through the code.  You'll notice that the average math & reading score are wildly off.  While trying to debug the code and find where the error was on 'per_school_passing_' I was getting an error on 'student_name' being undefined and decided to move on to the final project instead ☹️